pyproject.toml
[project]
name = "mmdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "python-fasthtml>=0.10.2",
]

[project.scripts]
mmdb = "mmdb.cli:main"